+++ src/main/java/dto/ArtistDTO.java
@@ -6,10 +6,10 @@ import java.util.Objects;
 
 public class ArtistDTO {
 
-    private final int id;
+    private final long id;
     private final String artist;
 
-    public ArtistDTO(int id, String artist) {
+    public ArtistDTO(long id, String artist) {
         this.id = id;
         this.artist = artist;
     }
@@ -19,7 +19,7 @@ public class ArtistDTO {
         this.artist = entity.getArtist();
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
