+++ src/main/java/dto/StatisticsDTO.java
@@ -5,23 +5,23 @@ import java.util.Map;
 
 public class StatisticsDTO {
 
-    private final Map<ArtistDTO, Integer> bestArtists;
-    private final Map<GenreDTO, Integer> bestGenres;
+    private final Map<ArtistDTO, Long> bestArtists;
+    private final Map<GenreDTO, Long> bestGenres;
     private final Map<LocalDateTime, String> abouts;
 
-    public StatisticsDTO(Map<ArtistDTO, Integer> bestArtists,
-                         Map<GenreDTO, Integer> bestGenres,
+    public StatisticsDTO(Map<ArtistDTO, Long> bestArtists,
+                         Map<GenreDTO, Long> bestGenres,
                          Map<LocalDateTime, String> abouts) {
         this.bestArtists = bestArtists;
         this.bestGenres = bestGenres;
         this.abouts = abouts;
     }
 
-    public Map<ArtistDTO, Integer> getBestArtists() {
+    public Map<ArtistDTO, Long> getBestArtists() {
         return bestArtists;
     }
 
-    public Map<GenreDTO, Integer> getBestGenres() {
+    public Map<GenreDTO, Long> getBestGenres() {
         return bestGenres;
     }
