+++ src/main/java/service/ArtistService.java
@@ -20,12 +20,12 @@ public class ArtistService implements IArtistService {
     }
 
     @Override
-    public boolean exists(int id) {
+    public boolean exists(long id) {
         return dataSource.exists(id);
     }
 
     @Override
-    public ArtistDTO get(int id) {
+    public ArtistDTO get(long id) {
         return dataSource.get(id);
     }
 
@@ -35,7 +35,7 @@ public class ArtistService implements IArtistService {
     }
 
     @Override
-    public void update(int id, String artist) {
+    public void update(long id, String artist) {
         if (dataSource.exists(id)) {
             dataSource.update(id, artist);
         } else {
@@ -44,7 +44,7 @@ public class ArtistService implements IArtistService {
     }
 
     @Override
-    public void delete(int id) {
+    public void delete(long id) {
         if (dataSource.exists(id)) {
             dataSource.delete(id);
         } else {
